Fixes an issue where a progress dialog could be shown with no content.
Once the WorkerProgressPane is given a worker, it can start its dialog at any time.  If the worker is already SCHEDULE or RUNNING, the dialog will be shown immediately, without content.  This change delays the call the set the WorkerProgressPane's worker until the dialog has all of it's content.

--HG--
branch : progress-dialog-fixes